refactor(alerting): decompose alarms_page + monitor_detail_flyout (ps48 H1/H2)
alarms_page.tsx (1216 → 763 LOC):
- New `useTimeRange` hook owns the picker state, sessionStorage round-trip,
  parse-failure self-healing, and stable callbacks. Exposes `bumpRefreshToken`
  for imperative refetch from the acknowledge flow.
- New `useDatasourceSelection` hook owns the priority-order resolver
  (localStorage → admin default → first discovered) and the persist effect.
- New `useRulesData` hook owns the rules fetch, loading/error/warnings,
  and exposes `refetch()` for the optimistic clone/save flows.
- New `AlarmsPageCallouts` sub-component renders the four-banner stack
  (alerting plugin missing, alerts error, page error, datasource warnings)
  the page used to inline.
- The page itself now reads as orchestration: hook calls + tab switch +
  flyout wiring. No reducer migration — functional `setState` already
  eliminates the stale-closure risks that motivate `useReducer`.

monitor_detail_flyout.tsx (1032 → 861 LOC):
- New `useMonitorDetail` hook owns the on-mount fetch + cancel-on-unmount.
- `ConditionPreviewGraph` (sparkline + sparse-data fallback) extracted to
  `monitor_detail/condition_preview_graph.tsx`.
- `humanizeCondition` (Painless → readable text) extracted to
  `monitor_detail/humanize_condition.tsx`.

No reducer rewrite, no behaviour change — purely structural extraction.
Existing alarms_page.test.tsx and monitor_detail_flyout.test.tsx pass
unmodified against the new shape.
